numExamplesUsed should be long (gets negative at epoch 7 on the full RNA-Seq training set).

+++ somatic/src/main/java/org/campagnelab/dl/framework/tools/TrainModel.java
@@ -258,7 +258,7 @@ public abstract class TrainModel<RecordType> extends ConditionRecordingTool<Trai
         System.out.println("errorEnrichment=" + args().errorEnrichment);
 
         performanceLogger.setCondition(args().experimentalCondition);
-        int numExamplesUsed = 0;
+        long numExamplesUsed = 0;
         int notImproved = 0;
 
         System.out.flush();
